Spot fix
I decreased the camera zoom on the x-axis to show off the animations for the turret during the Sunday meeting, since it awakens basically as it enters the frame on normal zoom.
However, I forgot to change it back before committing my changes, so I'm making this spot fix to change it back.