Hitbox Documentation, Interactables, Basic Dialogue, Camera Fix, EventBus
Hitbox, Hurtbox, and Attack have been given (mid) documentation and comments.

Interactable scene added, add it as a child of a node and it'll give you a little hitbox to potentially interact with (K on keyboard, Right Face Button on controller). These are purely for dialogue currently, but the class was made for the future.

Dialogue is functional, but does not have choice selectors at the moment.  Text is typed out over time which can be skipped. Dialogue text files are formatted as:

<TAG/COMMAND> <TEXT>
For example:
LINE Hello
EOF

LINE indicates a spoken line of dialogue. EOF tells the dialogue program to stop scanning (required at end of file). Currently only LINE and EOF are implemented, plans exist for choices (IF, ENDIF), branches based on player data (IFP, ENDIFP) and player state updates (PSTATE). These will likely require a bit of an overhaul.

Camera Smoothing was making movement jittery, it's disabled now.

EventBus is an auto-loaded file to place signals in that may be faster (or more possible) than relating two nodes.