Attacks and Knockbacks (part 1?)
Attack component is more fleshed out, and a slight rework of hitboxes has happened.

Hitboxes now scan when ".activate()"'d and have pierce functionality. If the "pierce" property is > -1, the hitbox will only hit the provided number of enemies plus one (zero pierce -> one hit).

This also comes with a change to Melee hitboxes in particular. They require a raycast childed to function, but now they'll detect if there's a body under "environment" and only hit things between said body and the hitboxes' parent.

If the player (who now has a 3-swing combo) only hits the environment, they'll be knocked back. Right now it's a bit clunky but it'll be fine-tuned at some point.

This applies to *all* melee hitboxes currently, so if you add one to something (and selfKnockback is set above 0) make sure it can take knockback. If you don't use the attack node, ensure the attacker has "apply_knockback" as a function. If you do use one,  make sure it has "take_knockback".